wip(spike): hoist session.jsonl parser to @lobu/core
Two HTTP surfaces parse the worker's `.openclaw/session.jsonl`:

- the worker's own `/session/messages` and `/session/stats` (rooted at
  `WORKSPACE_DIR`), and
- the gateway's `/session/messages` and `/session/stats` proxy fallback
  (rooted at `workspaces/<agentId>`, used when the worker is offline).

The gateway proxies to the worker when the worker is online and falls
back to its own copy otherwise — so the two parsers MUST agree on the
JSONL line shape. They had drifted. The worker used `safeJsonParse` (a
core util that debug-logs malformed lines); the gateway inlined a bare
`try { JSON.parse } catch {}`. The worker carried two extra
`SessionEntry` fields (`tokensBefore`, `firstKeptEntryId`) that no
parser actually read. The display-projection logic in `entryToMessage`
was duplicated character-for-character.

Move the shared parts (`SessionEntry`, `ParsedMessage`,
`parseSessionEntries`, `entryToMessage`) into
`packages/core/src/utils/session-file.ts` and have both call sites
import from `@lobu/core`. `safeJsonParse` wins as the canonical line
parser — it's already in core, and the debug-log on malformed lines is
strictly more useful than silently swallowing them.

What does NOT move: `findSessionFile`. The two call sites have
intentionally different path policies (worker scans its own
`WORKSPACE_DIR` one level deep; gateway scans `workspaces/<agentId>`
three levels deep with a `SAFE_AGENT_ID` regex guard against path
traversal). Collapsing them would force one side to inherit the other's
policy — not a behaviour change to make silently.

Dropped from the canonical `SessionEntry`: `tokensBefore` and
`firstKeptEntryId`. They were declared on the worker's local interface
but nothing reads them — only `agent-worker/src/__tests__/
memory-flush-runtime.test.ts` mentions them in a fixture. Reintroduce
when an actual consumer needs them.

Validation: `make build-packages` clean. `make typecheck` clean
(server + owletto). `bunx tsc --noEmit` clean inside `packages/agent-worker`.
