+++ packages/core/src/index.ts
@@ -154,6 +154,13 @@ export type { McpStatus, McpToolDef } from "./utils/mcp-tool-instructions";
 export * from "./utils/network-domains";
 export * from "./utils/retry";
 export * from "./utils/sanitize";
+// Shared OpenClaw session.jsonl parser (gateway + worker).
+export {
+  entryToMessage,
+  parseSessionEntries,
+  type ParsedMessage,
+  type SessionEntry,
+} from "./utils/session-file";
 export * from "./utils/urls";
 export * from "./worker/auth";
 export type {
