+++ packages/core/src/utils/session-file.ts
@@ -1,0 +1,141 @@
+/**
+ * Shared parser for OpenClaw `session.jsonl` files.
+ *
+ * Two HTTP surfaces read these files: the worker's `/session/messages` /
+ * `/session/stats` endpoints (rooted at the worker's own `WORKSPACE_DIR`)
+ * and the gateway's `/session/messages` / `/session/stats` REST endpoints
+ * (rooted at the gateway's `workspaces/<agentId>` tree, queried when the
+ * worker is offline). The gateway proxies to the worker when it's online
+ * and falls back to its own copy otherwise — so the two parsers must
+ * agree, and historically they had drifted (different fields kept on
+ * `SessionEntry`, different `JSON.parse` error handling, the same logic
+ * copy-pasted twice).
+ *
+ * Anything path-policy related (where to *look* for the file) stays at
+ * the call site — the worker scans one level under `WORKSPACE_DIR`; the
+ * gateway scans up to three levels under the per-agent workspace dir
+ * with a `SAFE_AGENT_ID` regex guarding the join. Those are intentionally
+ * different and must not be collapsed without an operator decision.
+ */
+
+import { safeJsonParse } from "./json";
+
+/**
+ * Raw entry shape as written to `session.jsonl` by the worker.
+ *
+ * `tokensBefore` / `firstKeptEntryId` (worker memory-flush bookkeeping)
+ * are not read by either parser today — left off this canonical shape on
+ * purpose; reintroduce when a consumer actually needs them.
+ */
+export interface SessionEntry {
+  type: string;
+  id: string;
+  parentId: string | null;
+  timestamp: string;
+  message?: {
+    role: string;
+    content: unknown;
+    usage?: { inputTokens?: number; outputTokens?: number };
+  };
+  summary?: string;
+  provider?: string;
+  modelId?: string;
+  customType?: string;
+  content?: unknown;
+  display?: boolean;
+}
+
+/** Display-friendly projection emitted to API consumers (`/session/messages`). */
+export interface ParsedMessage {
+  id: string;
+  type: string;
+  role?: string;
+  content: unknown;
+  model?: string;
+  timestamp: string;
+  isVerbose?: boolean;
+  usage?: { inputTokens?: number; outputTokens?: number };
+}
+
+/**
+ * Parse a session.jsonl blob into entries + the synthetic session id
+ * found on the leading `{type: "session", id}` line.
+ *
+ * - Splits on `\n` and skips blank lines (same as both pre-existing copies).
+ * - Uses {@link safeJsonParse} so malformed lines are skipped quietly with
+ *   a debug log (debug-only because production sessions occasionally
+ *   contain partial writes after crash/kill).
+ * - The leading `session` entry is extracted, not pushed into `entries`.
+ */
+export function parseSessionEntries(content: string): {
+  entries: SessionEntry[];
+  sessionId?: string;
+} {
+  const lines = content.split("\n").filter((l) => l.trim());
+  const entries: SessionEntry[] = [];
+  let sessionId: string | undefined;
+  for (const line of lines) {
+    const parsed = safeJsonParse<SessionEntry & { id: string }>(line);
+    if (!parsed) continue;
+    if (parsed.type === "session") {
+      sessionId = parsed.id;
+      continue;
+    }
+    entries.push(parsed);
+  }
+  return { entries, sessionId };
+}
+
+/**
+ * Project a single {@link SessionEntry} into the {@link ParsedMessage}
+ * display shape, or `null` for entry kinds that don't surface as
+ * user-visible messages (everything other than `message`, `compaction`,
+ * `model_change`, `custom_message`).
+ *
+ * `isVerbose` marks entries the UI hides behind a "verbose" toggle —
+ * tool results, compaction/model-change markers, custom system events
+ * that aren't explicitly displayed.
+ */
+export function entryToMessage(entry: SessionEntry): ParsedMessage | null {
+  if (entry.type === "message" && entry.message) {
+    return {
+      id: entry.id,
+      type: "message",
+      role: entry.message.role,
+      content: entry.message.content,
+      timestamp: entry.timestamp,
+      isVerbose: entry.message.role === "toolResult",
+      usage: entry.message.usage,
+    };
+  }
+  if (entry.type === "compaction") {
+    return {
+      id: entry.id,
+      type: "compaction",
+      content: entry.summary || "",
+      timestamp: entry.timestamp,
+      isVerbose: true,
+    };
+  }
+  if (entry.type === "model_change") {
+    return {
+      id: entry.id,
+      type: "model_change",
+      content: `${entry.provider}/${entry.modelId}`,
+      model: `${entry.provider}/${entry.modelId}`,
+      timestamp: entry.timestamp,
+      isVerbose: true,
+    };
+  }
+  if (entry.type === "custom_message") {
+    return {
+      id: entry.id,
+      type: "custom_message",
+      role: "user",
+      content: entry.content,
+      timestamp: entry.timestamp,
+      isVerbose: !entry.display,
+    };
+  }
+  return null;
+}
